fix hashCodes of models and move MainApp
