+++ main-service/src/main/java/ru/practicum/ewm/MainApp.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.main;
+package ru.practicum.ewm;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
