+++ main-service/src/main/java/ru/practicum/ewm/main/category/model/Category.java
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
+import java.util.Objects;
+
 @Data
 @NoArgsConstructor
 @Entity
@@ -28,6 +30,6 @@ public class Category {
 
     @Override
     public int hashCode() {
-        return getClass().hashCode();
+        return Objects.hash(id);
     }
 }
