+++ main-service/src/main/java/ru/practicum/ewm/main/user/model/User.java
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
+import java.util.Objects;
+
 @Data
 @NoArgsConstructor
 @Entity
@@ -31,6 +33,6 @@ public class User {
 
     @Override
     public int hashCode() {
-        return getClass().hashCode();
+        return Objects.hash(id);
     }
 }
